Практически закончена основная часть бота, не реализовано: команды, определение правильности входа по времени

+++ InformationProcessSupport.Data/Statistics/IStatisticRepository.cs
@@ -4,6 +4,6 @@
     {
         Task AddAsync(StatisticEntity statisticEntity);
         Task UpdateConnectionTimeAsync(StatisticEntity statisticEntity);
-        Task<int> GetStatisticIdByUserIdAndChannelId(ulong userId, ulong channelId);
+        Task<int> GetStatisticIdByUserIdAndChannelId(int userId, int channelId);
     }
 }
